ZADANIA/WWW_MySQL_Accountant/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
packaging==21.2
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
